worked on the menus and editor properties. - there is now a default for: style, indentation, line ending. - the three above can be set for the current file via the edit menu.
The utf-8 encoding stuff did not work because the Qscintilla text() method
assumes some encoding which is not utf-8 apparantly. Now I use SCO_GETTEXT with
sendScintilla to get the text (as utf- encoded bytes). This is done via a buffer.
For the getLine and getCurLine methods I need to figure out whether to produce
bytes or strings...