Fix issue #278: app nap on OSX 9.2 makes IEP kernel slow
Resolution is to use MinRK's appnap module that uses ctyopes
to talk to OSX and tell not to turn the process in a napping
state. We just take one module from the appnap package and
put in in the iepkernel subpackage. That module is only loaded
on OSX 9.2 and up.