run -i script.py
Add a menu entry to execute a script with %run -i.
Doing so is important because 
%load_ext autoreload
%autoreload 2
take effect. One can interactively change code and the changes takes effect. In the same function I added also the the change directory to script directory. The implementation is not optimal of course (I was not able to make shortcuts working). 
I think it would be very important to include this functionality in the IEP!!

See also following two discussions:
https://groups.google.com/forum/#!topic/pyzo/l3I_3AvrFdA
and
https://groups.google.com/forum/#!topic/pyzo/icfCtU_zUSw