History Viewer: Auto scroll fix (#526)
History Viewer now auto scrolls to the bottom of the list when new commands are entered in the shell instead of "getting stuck" at the beginning of the list, with newer commands disappearing out of the window. In addition, history viewer scrolls to the bottom automatically when starting the program.