implement scaled breakpoint visualization and remove cyclic references
This changes the size of the red breakpoint circles (and similar indicators
in the vertical breakpoint bar) proportional to the cursor height instead
of having a fixed pixel height.
To implement this, I had to re-implement the handling of all vertical bar
widths. This new implementation has the advantage that it removes the cyclic
references between editor and breakpoint classes.