remove limitations for usage of Python's logging module
This commit will only setup the logging configuration and add logging handlers when they are
required for post-mortem debugging, i.e. for the asyncio and old tornado event loop.
And even in that case, when Pyzo creates the basic logging configuration, it informs the user
about the force argument: logging.basicConfig(..., force=True). When that force argument is
used, Pyzo will reinstall the logging handlers so that post-mortem debuggin still works for
async code.