rework shell configuration dialog
Tab container is replaced with a stacked widget and a list.
The list selection switches the visible config form.
This is more comfortable when having many shell configurations.
Reordering via drag'n'drop is possible (as before with dragging tabs).
Duplicating of configurations via drag'n'drop is now possible.

The dialog window has now a larger default size so that everything can be seen without scrolling.

Some buttons had weird positions.
"Add config" was on top right in the tab bar widget
"Delete" and "Copy" were at the bottom of the QForm and not visible without scrolling or resizing of the dialog.
These buttons are now below the config list widget in the lower left corner of the dialog (and always visible).

Button "Copy" is renamed to "Duplicate".

Window resizing is now done before the dialog is shown so that there is no more flickering of the dialog window.

The "startupScript" text box has no more height limit.

Large text boxes "pythonPath", "startupScript" and "environ" now have QPlainTextEdit instead of QTextEdit so that there is no more undesired mixed formatting when pasting text.