improve vertical scrolling for search results in editor
old behavior:
When going to the next result that is outside the visible code lines, the editor
automatically scrolled to the correct place, but the cursor line was then at the
very top resp. very bottom of the editor.

new behavior:
When going to the next result that is outside the visible code lines, the editor will vertically center the line.